+++ lib/defopt.py
@@ -692,7 +692,11 @@ def _parse_docstring(doc):
     doc = str(NumpyDocstring(doc, cfg))
 
     with _sphinx_common_roles():
-        tree = docutils.core.publish_doctree(doc)
+        tree = docutils.core.publish_doctree(
+            # Disable syntax highlighting, as 1) pygments is not a dependency
+            # 2) we don't render with colors and 3) SH breaks the assumption
+            # that literal blocks contain a single text element.
+            doc, settings_overrides={'syntax_highlight': 'none'})
 
     class Visitor(NodeVisitor):
         optional = [